apzkr-pzpi-21-2-klovanyk-stepan/Task1-Server/requirements.txt
Flask==3.0.3
flask_bcrypt==1.0.1
Flask_Cors==4.0.1
flask_login==0.6.3
flask_sqlalchemy==3.1.1
icecream==2.1.3
marshmallow==3.21.3
marshmallow_sqlalchemy==1.1.0
Pillow==10.4.0
SQLAlchemy==2.0.18
SQLAlchemy_Utils==0.41.2
Werkzeug==3.0.3
